CI: Do not test against rails master branch
We don't need this on every CI run. We don't ever look at it,
so it's kind of a waste of electricity.

We can always run it locally if we need to:

BUNDLE_GEMFILE=test/gemfiles/Gemfile.rails-master bundle exec rake

We also don't have enough volunteers to comb through it when it
fails and submit bug reports. If someone wants to do that, let
me know and we can add it back into the CI run.