Delete AuthlogicLoadedTooLateError
This error no longer seems to be necessary.

I tested this in a new rails 6.0.0 application. In the following,
my debug output is prefixed with "->".

```
bin/rails server
-> loading rails_adapter.rb
=> Booting Puma
...
* Listening on tcp://localhost:3000
...
-> RailsImplementation included into ActionController::Base
-> RailsImplementation included into ActionController::API
-> loading application_controller.rb
Started GET "/" for 127.0.0.1 at 2019-09-11 13:33:05 -0400
Processing by HomeController#index as HTML
-> activate_authlogic
  Rendering home/index.html.erb within layouts/application
```

We can see that the `activate_authlogic` callback still occurs. So,
raising an AuthlogicLoadedTooLateError would be incorrect.

And, of course, I tested a normal sign in POST.